+++ sandstone/pattern-virtualgridlist-infinite-scroll/README.md
@@ -1,6 +1,6 @@
 ## VirtualGridList With Infinite Scroll Pattern
 
-A sample Enact application that shows off how to load items with infinite scroll in VirtualGridList
+A sample Enact application that shows off how to load items with infinite scroll in VirtualGridList.
 
 Run `npm install` then `npm run serve` to have the app running on [http://localhost:8080](http://localhost:8080), where you can view it in your browser.
 
@@ -10,7 +10,7 @@ Run `npm install` then `npm run serve` to have the app running on [http://localh
 - `sandstone/Panels/Panel`
 - `sandstone/VirtualList/VirtualGridList`
 
-In this app, Load data when scrolling reaches the bottom of the virtualgridlist. Use your API server to data fetch to load the data.
+In this app, load data when scrolling reaches the bottom of the virtualgridlist. Use your API server to data fetch to load the data.
 
 ---
